week_2_listening_comp/requirements.txt
# AWS and API Dependencies
boto3>=1.34.0
youtube-transcript-api>=0.6.2
chromadb>=0.4.22

# Web Framework
streamlit>=1.29.0
requests>=2.31.0

# Audio Processing
pydub>=0.25.1
soundfile>=0.12.1
ffmpeg-python>=0.2.0
openai-whisper==20231117

# Chinese Language Processing
pypinyin>=0.49.0
jieba>=0.42.1
zhon>=2.0.2
deep-translator>=1.11.4

# Data Processing
pandas>=2.1.4
numpy>=1.26.2

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.1
regex>=2023.10.3

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
